Secure the views for only admin management

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -116,8 +117,13 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """Add product to the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry you are not an admin, you can not do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -125,7 +131,8 @@ def add_product(request):
             messages.success(request, 'You added a product with success')
             return redirect(reverse('product_detail', args=[product.id]))
         else:
-            messages.error(request, 'Adding product failes. Please ensure the form is valid.')
+            messages.error(request, 'Adding product failes. \
+                Please ensure the form is valid.')
     else:
         form = ProductForm()
 
@@ -136,8 +143,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """Edit a product in the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry you are not an admin, you can not do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -146,7 +158,8 @@ def edit_product(request, product_id):
             messages.success(request, 'Product updated!')
             return redirect(reverse('product_detail', args=[product.id]))
         else:
-            messages.error(request, 'Updating product failed. Please ensure the form is valid.')
+            messages.error(request, 'Updating product failed. \
+                Please ensure the form is valid.')
     else:
         form = ProductForm(instance=product)
         messages.info(request, f'You are editing {product.name}')
@@ -159,8 +172,12 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """Delete a product from the store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry you are not an admin, you can not do that.')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Products deleted!')
